Merge dns-bday-pdox-3347
Author: exarkun
Reviewer: foom
Fixes: #3347

Change `twisted.names.client.Resolver` so that it will not put
multiple requests for the same record set onto the network at
once.  Instead, if it is asked for a record set while it is
waiting for a response from the network which will give that
record set, it will just make sure to deliver those results
to both callers once they arrive.

This is primarily beneficial because it avoids exposing the
UDP DNS client to a "birthday paradox" attack, where it can
be easier to successfully spoof a response by causing it to
issue many duplicates of a query with different message IDs.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@24305 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb