Merge delete-problemsFromTransport-2874
Author: exarkun
Reviewer: therve
Fixes: #2874

Remove the private `problemsFromTransport` API which exposed OpenSSL errors.
This API was redundant with `IProtocol.connectionLost`, which is called
with a `Failure` wrapped around an `OpenSSL.SSL.Error` if there is a problem
with SSL which results in the connection being lost.

AMP is also changed to use the failure passed to `connectionLost` rather
than `problemsFromTransport`.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@24403 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb