Merge web-wsgi-2753
Author: exarkun
Reviewer: therve
Fixes: #2753

Add a rudamentary WSGI container to `twisted.web`.

`twisted.web.http.Request.setResponseCode` is also changed to be more
strict in the values it accepts, raising a `TypeError` immediately if
given a non-integer for the status code.  Previously this would only
raise an exception later when an attempt was made to log the request.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@24584 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb