Merge showwarning-line-3211
Author: exarkun
Reviewer: therve
Fixes: #3211

Change the `warnings.showwarning` hook installed by `twisted.python.log`
so that it accepts a new, optional `line` parameter which is passed by
the Python warning system in Python 2.6.  This silences a huge number of
warnings when using Python 2.6 which are emitted when using any Twisted
code which invokes the Python warning system.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@24858 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb