Merge distrib-wantpid-2295-2
Author: exarkun
Reviewer: ralphm
Fixes: #2295

Add an optional parameter to the AF_FILE style of strport
description, `lockfile`.  By default, a "unix:" strport
will attempt to acquire a lock associate with the socket
specified.  If `lockfile=0` is given, this behavior will
be disabled.

`twisted.web.distrib` is also changed to allow a strports
description for the port it listens on, allowing a lock
file to be used or not with the port it listens on.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@24991 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb