Merge directory-lister-children-3631
Author: exarkun
Reviewer: therve
Fixes: #3631

Fix `DirectoryLister` so that it allows `Resource` (its base class) to properly initialize
itself.  This fixes an `AttributeError` raised when looking up a child of `DirectoryLister`.

Additionally, some of the tests have been changed to use `resource.getChildForRequest` rather
than calling `getChild` directly, as the former is a more realistic approximation of what
happens in a real application.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@26221 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb