Merge mutate-users-3735
Author: exarkun
Reviewer: glyph
Fixes: #3735

Fix the web-facing finger tutorial interface so that it notices updates to the
user file.  Previously user data was stored in a newly created dictionary once
in a while and the web server was never made aware of this new dictionary.  Now
the same dictionary is re-used each time user data is reloaded.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@26709 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb