Merge spawnprocess-file-mode-3567
Author: abbeyj, exarkun
Reviewer: therve, jesstess
Fixes: #3567

Executing a "script" on Windows (i.e., a file which starts with a #!
line) in Python 2.6 now works after a change to read the script file with the
real mode "rU" instead of the non-mode "ru". In earlier versions of Python on
Windows, unknown mode letters were ignored.



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@27356 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb