Merge loopingcall-3671
Author: tenth, glyph, exarkun
Reviewer: glyph, therve, ralphm, exarkun, washort

Add `LoopingCall.skipCounts`, a new `LoopingCall` constructor which accepts a
one-argument callable and invokes it on each interval with the number of
intervals which have elapsed since the last call.  This allows `LoopingCall`
to be used in "frame" contexts - eg, to redraw graphics, or update a simulation,
etc - and account for and respond to scheduling inadequacies by making up for
the missed iterations.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@27402 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb