Merge lore-lint-getattribute-fixes-4115
Author: exarkun
Reviewer: therve
Fixes: #4115

Add a test for one of the lore lint checks related to anchor tags and
fix the implementation so that it works with minidom's getAttribute
method (which only accepts one argument).  Also delete another of the
anchor tag lint checks, the one which looked for license text being
included in source listings, as it seems somewhat useless.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@27543 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb