Merge full-child-environment-4203
Author: exarkun
Reviewer: moshez
Fixes: #4203

For the Twisted Conch Insults tests which interact with a child process, ensure that
the child process is set up with an environment more closely resembling the parent,
in case any of its contents are necessary for proper operation.  In particular, this
fixes these tests on PyPy, which currently requires objdump to be somewhere on PATH.

Also account for some minor differences in the way certain exceptions behave between
CPython and PyPy in these tests.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@28027 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb